+++ dds/DdsDcpsCore.idl
@@ -79,22 +79,27 @@ module DDS {
   // are all part of the security specification (ptc/2017-09-26).
   // These structs/typedefs are in this file to satisfy dependencies for
   // DomainParticipantQos (which was extended in the security specification).
+  @OpenDDS::internal::special_serialization
   struct Property_t {
     string name;
     string value;
     boolean propagate;
   };
+  @OpenDDS::internal::special_serialization("prop_seq")
   typedef sequence<Property_t> PropertySeq;
 
+  @OpenDDS::internal::special_serialization
   struct BinaryProperty_t {
     string name;
     OctetSeq value;
     boolean propagate;
   };
+  @OpenDDS::internal::special_serialization("prop_seq")
   typedef sequence<BinaryProperty_t> BinaryPropertySeq;
 
   // This is for when we support XCDR1.
   // @appendable
+  @OpenDDS::internal::special_serialization
   struct PropertyQosPolicy {
     PropertySeq value;
     BinaryPropertySeq binary_value;
